requirements/local.txt
asgiref==3.7.2
coverage==7.3.2
Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==20.1.0
fontawesomefree==6.4.2
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3